SSR: Extract error code out to a separate module
This is to make reusing it outside of parsing easier in a subsequent
change.
